requirements.txt
absl-py
optax~=0.1.4
dm-haiku
numpy~=1.24.0
jax~=0.4.1
jaxlib
tensorflow-datasets
tensorflow
jupyter
matplotlib
tensorflow-gpu~=2.10.1